Add 'chatonly' configuration
The default 'monitor-activity' behavior in tmux will show activity when
something changes on the screen.  This will get triggered by multiple
things, like the ping time update, a change in the number of total or
active users, or signon-signoff-nickchange messages.  This change adds a
'chatonly' configuration option that supresses all of the above, thus
ensuring that tmux will only see activity that is actual chat traffic.

Closes: #19